changed the music screen pattern, from widgets to a new music.dart file, creating a class of it's own
